+++ src/components/MyProfile.js
@@ -1,18 +1,10 @@
-import { useSelector } from 'react-redux';
+import React from 'react';
+import MissionProfile from './missionComponents/MissionProfile';
 
-const MyProfile = () => {
-  const missions = useSelector((state) => state.missionsReducer);
-  const joinedMissions = missions.filter((mission) => mission.reserved === true);
-  return (
-    <div className="profile-missions">
-      <h2>My Missions</h2>
-      <ul className="missions-list">
-        {joinedMissions.map((mission) => (
-          <li className="joined-mission" key={mission.id}>{mission.name}</li>
-        ))}
-      </ul>
-    </div>
-  );
-};
+const MyProfile = () => (
+  <div className="profile">
+    <MissionProfile />
+  </div>
+);
 
 export default MyProfile;
