+++ src/components/styling/Header.css
@@ -1,8 +1,3 @@
-body {
-  padding: 0;
-  margin: 0;
-}
-
 header {
   display: flex;
   justify-content: space-between;
@@ -13,7 +8,7 @@ header {
 }
 
 .logo-corner {
-  width: 40%;
+  width: 60%;
   display: flex;
 }
 
@@ -29,18 +24,19 @@ header {
 
 nav {
   padding-right: 5%;
-  width: 25%;
+  width: 35%;
 }
 
 .nav-menu {
   list-style: none;
   width: 100%;
   display: flex;
-  justify-content: space-between;
   align-items: center;
+  gap: 20%;
 }
 
 .nav-items > a {
+  width: 30%;
   text-decoration: none;
   color: #047dff;
 }
@@ -48,4 +44,12 @@ nav {
 .active {
   text-decoration: underline;
   font-weight: 700;
+}
+
+@media (max-width: 992px) {
+  .logo {
+    margin: 0.35rem;
+    font-size: 1.5rem;
+    line-height: 2rem;
+  }
 }
